Notification moderator: email of commentator added. Link as is but request to log in for moderating comment added.
New moderator notification with email and link to commented page and request to login.
No links to approve and delete: due to CSRF direct links to modification of Plone objects result in request to confirm. So page with comments is presented to moderator.
If already logged in, moderator is on comment to moderate.
If not logged in, moderator is on login page with came_from.