When installing plone.app.discussion enable comments globally.
In 6.1 you have to explicitly install the add-on in the Add-ons control panel.
If you do that, then you obviously want discussions to be enabled globally.
I think this also helps for the Volto case, where a Discussion control panel is missing, so in Plone 6.0 you cannot even properly enable comments.